Configure modern-looking Windows LAF tabs. This completes NETBEANS-5673.
The old tab UI implementations are still left in a working state, and can be
reverted to by changing "WinFlatEditorTabDisplayerUI" to
"Windows8VectorEditorTabDisplayerUI" and "WinFlatViewTabDisplayerUI" to
"Windows8VectorViewTabDisplayerUI" in Windows8LFCustoms.

The Windows LAF's 'close tab' icon is adjusted to be vertically symmetrical.
This change works well with either the old or the new tab styles.